test(db,integration): Remove unnecessary to_string calls and improve assertions
Update all database and integration tests to use string literals and
improve assertion patterns for better test quality.

Database Tests (~30 files):
- orm/tests/*.rs: Remove .to_string() in query tests
- hybrid/tests/*.rs: Use static strings in hybrid query tests
- contenttypes/tests/*.rs: Improve content type registry tests
- backends/tests/*.rs: Clean up backend connection tests
- pool/tests/*.rs: Improve connection pool tests

Integration Tests (~100+ files):
- admin/*.rs: Admin panel integration tests
- api/*.rs: REST API integration tests
- auth/*.rs: Authentication tests
- contrib/*.rs: Contrib module tests
- di/*.rs: Dependency injection tests
- messages/*.rs: Message framework tests
- middleware/*.rs: Middleware tests
- orm/*.rs: ORM integration tests
- pagination/*.rs: Pagination tests
- rendering/*.rs: Template rendering tests
- rest/*.rs: REST framework tests
- routing/*.rs: URL routing tests
- server/*.rs: Server tests
- sessions/*.rs: Session tests
- settings/*.rs: Settings system tests
- utils/*.rs: Utility tests
- validation/*.rs: Validation tests
- viewsets/*.rs: ViewSet tests

Pattern Changes:
```rust
// Before
let name = "test".to_string();
assert_eq!(result, true);
User { username: "alice".to_string(), ... }

// After
let name = "test";
assert!(result);
User { username: "alice", ... }
```

Benefits:
- Faster test execution (fewer allocations)
- Cleaner test code
- Consistent patterns across test suite
- Improved readability

Rationale: Tests should follow the same quality standards as production
code, using static strings where possible and idiomatic assertions.

+++ tests/integration/tests/pagination/cursor_integration.rs
@@ -94,7 +94,7 @@ async fn setup_test_db(
 		.expect("Failed to insert test article");
 
 		articles.push(TestArticle {
-			id: i as i32,
+			id: i,
 			title,
 			content,
 			author_id,
